requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.7
django-appconf==1.0.6
django-cors-headers==4.3.1
django-filter==23.4
django-phonenumber-field==7.2.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
phonenumbers==8.13.26
pilkit==3.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
pycryptodome==3.19.0
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
